Refactor to vulns
