+++ internal/utility/vulns/vulnerability.go
@@ -1,4 +1,4 @@
-package modelsutil
+package vulns
 
 import (
 	"fmt"
