+++ internal/utility/vulns/vulnerability_test.go
@@ -1,10 +1,10 @@
-package modelsutil_test
+package vulns_test
 
 import (
 	"testing"
 	"time"
 
-	"github.com/google/osv-scanner/internal/modelsutil"
+	"github.com/google/osv-scanner/internal/utility/vulns"
 	"github.com/google/osv-scanner/pkg/lockfile"
 	"github.com/google/osv-scanner/pkg/models"
 )
@@ -19,7 +19,7 @@ func expectIsAffected(t *testing.T, vuln models.Vulnerability, version string, e
 		CompareAs: lockfile.NpmEcosystem,
 	}
 
-	if modelsutil.IsAffected(vuln, pkg) != expectAffected {
+	if vulns.IsAffected(vuln, pkg) != expectAffected {
 		if expectAffected {
 			t.Errorf("Expected OSV to affect package version %s but it did not", version)
 		} else {
@@ -95,7 +95,7 @@ func TestOSV_AffectsEcosystem(t *testing.T) {
 			Affected:  test.Affected,
 		}
 
-		if modelsutil.AffectsEcosystem(vuln, test.Ecosystem) != test.Expected {
+		if vulns.AffectsEcosystem(vuln, test.Ecosystem) != test.Expected {
 			t.Errorf(
 				"Test #%d: Expected OSV to return %t but it returned %t",
 				i,
@@ -114,7 +114,7 @@ func TestOSV_AffectsEcosystem(t *testing.T) {
 		Affected:  nil,
 	}
 
-	if modelsutil.AffectsEcosystem(vuln, "npm") {
+	if vulns.AffectsEcosystem(vuln, "npm") {
 		t.Errorf(
 			"Expected OSV to report 'false' when it doesn't have an Affected, but it reported true!",
 		)
